requirements.txt
openai
PyYAML
pytest